backend/requirements.txt
# ── ML ────────────────────────────────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
ultralytics>=8.1.0
timm>=0.9.16
open-clip-torch>=2.24.0

# ── CV / Image ────────────────────────────────────────────────────────────────
Pillow>=10.1.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
scipy>=1.12.0
scikit-image>=0.22.0
albumentations>=1.4.0

# ── Geospatial ────────────────────────────────────────────────────────────────
rasterio>=1.3.9
pyproj>=3.6.1
shapely>=2.0.3
geojson>=3.1.0

# ── API ───────────────────────────────────────────────────────────────────────
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ── Observability ─────────────────────────────────────────────────────────────
loguru>=0.7.2
prometheus-fastapi-instrumentator>=6.4.0

# ── Utilities ─────────────────────────────────────────────────────────────────
httpx>=0.26.0
python-dotenv>=1.0.1
huggingface-hub>=0.20.0
tqdm>=4.66.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# ── Dev ───────────────────────────────────────────────────────────────────────
ruff>=0.2.0
black>=24.1.0
